requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-milvus==0.1.9
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.28
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.72.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
posthog==3.23.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pulsar-client==3.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pymilvus==2.5.6
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.32.3
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
